refactor(functions): extract minimal store
